Commit before changing configuration settings approach

+++ src/main/java/devidin/net/yavumeter/display/Displayer.java
@@ -2,13 +2,13 @@ package devidin.net.yavumeter.display;
 
 public abstract interface Displayer {
 	
-	public abstract void displayLR(int[] amplitudeLR);
-	public abstract void displayLRasNumber(int[] amplitudeLR);
-	public abstract void displayLRasNumbers(int[] amplitudeLR);
+	public abstract void displayLR(double[] amplitudeLR);
+	public abstract void displayLRasNumber(double[] amplitudeLR);
+	public abstract void displayLRasNumbers(double[] amplitudeLR);
 	
-	public abstract void display(int[] amplitudeLR, int channels);
-	public abstract void displayAsNumber(int[] amplitudeLR, int channels);
-	public abstract void displayAsNumbers(int[] amplitudeLR, int channels);
+	public abstract void display(double[] amplitudeLR, int channels);
+	public abstract void displayAsNumber(double[] amplitudeLR, int channels);
+	public abstract void displayAsNumbers(double[] amplitudeLR, int channels);
 	
 	public abstract void init();
 	public abstract void shutdown();
